Cleaned

+++ persistence/src/main/java/com/softserve/osbb/repository/UserRepository.java
@@ -14,12 +14,11 @@ import java.util.List;
 @Repository
 public interface UserRepository extends JpaRepository<User,Integer> {
 
-    public User findByUserId(long id);
+    public User findByUserId(Integer id);
     public List<User> findUserByEmail(String emailAddress);
     public List<User> findByFirstName(String firstName);
     public List<User> findByLastName(String lastName);
     public List<User> findByRole(Role role);
     public List<User> findByOsbb(Osbb osbb);
-
 
 }
